+++ CHANGELOG.md
@@ -5,6 +5,13 @@ BUG FIXES:
 * resources/opennebula_user: Fix ignored renaming (#343)
 * resources/opennebula_group: Fix ignored renaming (#343)
 
+ENHANCEMENTS:
+
+* resources/opennebula_group_admins: Replace Typelist by Typeset on `users_ids`
+* resources/opennebula_user: Replace Typelist by Typeset on `groups`
+* resources/opennebula_virtual_data_center: Replace Typelist by Typeset on `group_ids`, `host_ids`, `datastore_ids`, `vnet_ids`, `cluster_ids`
+* resources/opennebula_network: Replace Typelist by Typeset on `clusters`, `security_groups`
+
 ## 1.0.0 (September 19th, 2022)
 
 BUG FIXES: